flexible provider registration
This commit decouples the dns provider implementation from the generic
dns package. All coding related to a dedicated provider is bundled in
the provider package (including the validation coding for the validating
webhook). This is achieved by introducing a ProviderFactory
interface and a registry (replacing the dnsProviderConstructors struct)
based on this interface.

Additionally a ne generic configuration (fields Kind and ProviderConfig)
have been introduced for the provider configuration. It can be used instead
of the former dedicated provider specific fields. To be compatible
both mechanisms can still be used.

This change offers a better flexibility for adding new providers.
A new provider can just be added by registering its ProviderFactory.
There is no code change required in the rest of the component.

The registry still allows mocking of providers by using a dedicated
instance for testing filled with the required provider mocks.